+++ lib/commands/sudo.dart
@@ -1,0 +1,1 @@
+// https://pub.dev/packages/shell