+++ lib/dshell.dart
@@ -1,0 +1,18 @@
+export 'commands/cat.dart' show cat;
+export 'commands/cd.dart' show cd;
+export 'commands/copy.dart' show copy;
+export 'commands/delete.dart' show delete;
+export 'commands/echo.dart' show echo;
+export 'commands/fileList.dart' show fileList;
+export 'commands/find.dart' show find;
+export 'commands/is.dart' show isFile, isDirectory, exists;
+export 'commands/make_dir.dart' show makeDir;
+export 'commands/move.dart' show move;
+export 'commands/pop.dart' show pop;
+export 'commands/push.dart' show push;
+export 'commands/pwd.dart' show pwd;
+export 'commands/read.dart' show read;
+export 'commands/sleep.dart' show sleep;
+export 'commands/settings.dart' show Settings;
+export 'commands/sudo.dart';
+export 'commands/touch.dart' show touch;