+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.3
+ Documentation formatting.
+
+## 1.0.2
+ Tweaks to the documenation.
+
+ Provided an example.dart
+
 ## 1.0.1
 
 - Initial release.